+++ src/cljs/withings_client/core.cljs
@@ -1,7 +1,7 @@
 (ns withings-client.core
   (:require
    [ajax.core :refer [GET POST]]
-   [cljs.core.async :refer [<!]]
+   ;; [cljs.core.async :refer [<!]]
    [clojure.string :as string]
    [goog.events :as events]
    [goog.history.EventType :as HistoryEventType]
@@ -10,12 +10,17 @@
    [reagent.dom :as rdom]
    [reitit.core :as reitit]
    [withings-client.ajax :as ajax])
-  (:require-macros
-   [cljs.core.async.macros :refer [go]])
+  #_(:require-macros
+     [cljs.core.async.macros :refer [go]])
   (:import
    goog.History))
 
-(defonce session (r/atom {:page :home}))
+(defonce session (r/atom {:page :home
+                          :name nil
+                          :cid nil
+                          :secret nil
+                          :belong nil
+                          :email nil}))
 
 (defn nav-link [uri title page]
   [:a.navbar-item
@@ -45,12 +50,15 @@
    ;; test csrf-token
    [:p js/csrfToken]])
 
-;;
-(def redirect-uri "https://wc.melt.kyutech.ac.jp/callback")
-(def withings-uri "https://account.withings.com/oauth2_user/authorize2")
+
+(def redirect-uri js/redirectUrl)
+
 (def scope "user.metrics,user.activity,user.info")
+
+(def authorize2-uri "https://account.withings.com/oauth2_user/authorize2")
+
 (def base
-  (str withings-uri
+  (str authorize2-uri
        "?response_type=code&redirect_uri=" redirect-uri "&"
        "scope=" scope "&"))
 
@@ -59,6 +67,7 @@
   (str base "client_id=" (:cid @session) "&state=" (:name @session)))
 
 (defn create-user!
+  ":name, :cid, :secret は必須フィールド。元バージョンはチェックが抜けている。"
   [params]
   (POST "/api/user"
     {:format :json
@@ -67,26 +76,25 @@
       "x-csrf-token" js/csrfToken}
      :params params
      :handler (fn [_] (js/alert (str "OK " params)))
-     :error-handler (fn [e] (.log js/console (str e)))}))
-
+     :error-handler (fn [e] (js/alert (str  "error " e)))}))
 
 (defn new-component []
   [:div
    [:h2 "new"]
-   [:div [:label {:class "label"} "name"]]
+   [:div [:label {:class "label"} "name(*)"]]
    [:div {:class "field"}
     [:input {:value (:name @session)
              :on-change #(swap! session
-                               assoc
-                               :name
-                               (-> % .-target .-value))}]]
-   [:div [:label {:class "label"} "cid"]]
+                                assoc
+                                :name
+                                (-> % .-target .-value))}]]
+   [:div [:label {:class "label"} "cid(*)"]]
    [:div {:class "field"}
     [:input {:on-change #(swap! session
                                 assoc
                                 :cid
                                 (-> % .-target .-value))}]]
-   [:div [:label {:class "label"} "secret"]]
+   [:div [:label {:class "label"} "secret(*)"]]
    [:div {:class "field"}
     [:input {:on-change #(swap! session
                                 assoc
@@ -106,20 +114,21 @@
                                 (-> % .-target .-value))}]]
    [:div {:class "field"}
     [:button {:class "button is-primary is-small"
-              :on-click #(do
-                           (create-user!
-                            (select-keys @session
-                                         [:name :cid :secret :belong :email]))
+              :on-click #(let [params (select-keys
+                                       @session
+                                       [:name :cid :secret :belong :email])]
+                           (create-user! params)
                            (swap! session
                                   assoc
-                                  :uri (create-url)))}
+                                  :uri
+                                  (create-url)))}
      "create"]]])
 
 (defn link-component []
   [:div
-   [:p "create ボタンの後、下に現れるリンクをクリック。" [:br]
-    "create のタイミングで name, belong, email を DB インサートするため、"
-    "create を省略できない。"]
+   [:p "(*)は必須フィールド。belong, email はカラでもよい。" [:br]
+    "create ボタンの後、下に現れるリンクをクリックし、"
+    "acccess トークン、refresh トークンを取得する。"]
    [:p "クリックで登録 → " [:a {:href (:uri @session)} (:name @session)]]])
 
 (defn users-component []